remove obsolete (plugin)repositories
After I played around with the SAT plugin in some example projects, I
realized that artifacts are requested by JCenter without having the
repository enabled in the artifacts itself.
I don't think it is good to use a plugin that injects additional
repositories into the build.
After removing the repositories and plugin respitories in the parent POM
of SAT, I tried to build it using an empty local Maven repo. It
succeeded, so IMHO that third party repositories are not necessary
anymore.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>