Remove MavenPomderivedInClasspathCheck
This check only results in false positives now a POM first approach with m2e in Eclipse is used.
m2e uses maven.pomderived for keeping track of classpath entries that are recreated during project import and when updating the project configuration.
The check is already suppressed in many repositories so it's better to stop using it.

Signed-off-by: Wouter Born <github@maindrain.net>